Move delivery_carrier_label_postlogistics from __unported__ to root folder
Set delivery_carrier_label_postlogistics installable

Extract the description in README.rst

New API for res.company

Migration of delivery.py to the new API

The state field has been replaced by boolean fields

Migration of stock picking to the new API

And API changes in base_delivery_carrier_label

compute method must assign a recordset, not a list

Wrong assignation

Migration of res.config to new API

Fix a comment

Comma at end of line

Migrated base_delivery_carrier_label so adapt the call

Migrated here: https://github.com/bealdav/carrier-delivery/pull/1

The field on the package is now parcel_tracking

The relation is now package_id

The domain prevents other modules to show their options

The drawback of this change is that postlogistics will also
show the options of the others carriers but I can't find a way
to filter them out. The user still see in the search that an option
is not related to postlogistics.

Fallback on 'en' when no lang is provided

Update the carrier inside the loop

Get the allowed options on the parent instead of the related

The related allowed_option_ids is computed before carrier_id is set and is
incorrect. Also, a change needs to be saved to see the change back on the
related field.

Hide the option of other carriers such as colissimo

The domain in the view is insufficient because it displays the options with an
empty partner.  Adding dynamically the domain when carrier's partner is
postlogistics ensures that only postlogistics' options are shown.

Fix: W503 line break before binary operator

Readme more similar than the OCA's template

Add license key and OCA logo

Remove dead code

Implement Postlogistics Cash on Delivery labels

The webservice awaits the cash on delivery amount.
When a whole sales order is delivered in one picking / one package, we the
amount is the total of the sales order. If it is split in several pickings /
packages, we don't know the value of each package and so we ask the user to
fill the price on the packages.

Update translations

Fix error when user has no language

Set default language to 'en_US' when the user has no language