[FIX] Remove occurences of carrier type
Field carrier_type was originally type and was used to differentiate delivery carriers.
However odoo did implement delivery_type since v9 with this exact goal :
https://github.com/odoo/odoo/blob/11.0/addons/delivery/models/delivery_carrier.py#L22
This is also the field used in the implementation of enterprise modules for carriers.
Therefore we should use it instead of redefining a new one..