Improve handling of free fee SO lines removal for 13.0
Description of the issue on #306.
In short: the delivery sale line is required in 13.0 to be able:

* see which delivery method is used
* change the delivery method or update its price
* remove the delivery method

Removing entirely the line when the fee is free is therefore an issue.
The free delivery line is kept, although greyed, in the sale order.
However, it is considered as non-invoiceable, so it doesn't create an
invoice line, and it is hidden from the sale report.

Closes #306

[UPD] Update delivery_free_fee_removal.pot

[UPD] README.rst

delivery_free_fee_removal 13.0.1.0.1

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: delivery-carrier-13.0/delivery-carrier-13.0-delivery_free_fee_removal
Translate-URL: https://translation.odoo-community.org/projects/delivery-carrier-13-0/delivery-carrier-13-0-delivery_free_fee_removal/