Improve handling of free fee SO lines removal for 13.0
Description of the issue on #306.
In short: the delivery sale line is required in 13.0 to be able:

* see which delivery method is used
* change the delivery method or update its price
* remove the delivery method

Removing entirely the line when the fee is free is therefore an issue.
The free delivery line is kept, although greyed, in the sale order.
However, it is considered as non-invoiceable, so it doesn't create an
invoice line, and it is hidden from the sale report.

Closes #306