[ADD] Send delivery notification from any operation
New module 'delivery_send_to_shipper_at_operation' allowing to send the
delivery notification from any operation (among chained operations) before
the delivery.
With a pick/pack/ship setup this allows to send the delivery notification to
the carrier when the pack operation is marked as done for instance
(by default Odoo sends it on the ship/delivery).

Once the notification has been sent from a non-delivery operation, it won't
be sent again when the delivery operation is validated.

Drawback:

This way of sending delivery notifications is not compatible if we
have several pack operations to process before the delivery (backorders?).
In such a case, the notification will be generated from only part of the
goods to deliver, which is not expected.