[FIX] delivery_multi_destination: Correctly handle base_on_rule subcarriers
Depending on `tentative_carrier` here is hacky beyond belief,
and there should exist a better way.

The problem this solves is as follows: Although `send_shipping()` is
called using the subcarrier, some code in
`base_on_rule_send_shipping()` (if the subcarrier has delivery type
`base_on_rule`) uses the _picking_'s carrier to call
`_get_price_available()`. This obviously breaks, because the picking's
carrier is the subcarrier's parent, and has no rules configured.

So we need to figure out what the chosen subcarrier was. The problem
is that there exists no hypothetical method
`determine_chosen_subcarrier_for_sale_order()` in this module.
Instead, the chosen subcarrier is determined by simply trying to call
`send_shipping()` until you find one that works.

So either we write the hypothetical
`determine_chosen_subcarrier_for_sale_order()` method and use it in
both `send_picking()` and here, or we use this workaround hack of
storing `tentative_carrier` on the order. But because you're reading
this, you can see which option was chosen.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>