[IMP] delivery_cttexpress: force delivery_price_method dependency
CTT Express doesn't have a delivery prices methods in their API. Until
now, it was enforced in the instructions to add delivery_price_method to
the instance in order to be able to have price rules. We force the
dependecy so the module functionality is complete out of the box.

TT40759