[FIX] delivery_auto_refresh: Don't fail on fixed rate carriers
Steps to reproduce:

- Create a delivery method (delivery.carrier) with a fixed rate.
- Enable the system parameter "delivery_auto_refresh.refresh_after_picking".
- Create a sales order with such delivery method and storable products.
- Confirm the sales order.
- Validate the picking.

Result: It can't be validated with the message "There is no matching
delivery rule.".

Explanation: Previous code computes the shipping values and call
directly to `_get_price_from_picking`, which uses rules and doesn't
take into account the rest of possibilities (integration or fixed
price).

We keep such computation for carriers based on rules, as they are not
properly updated, but added the possibility of writing back the price
for other carriers with fixed price or integrations, using the
`carrier_price` field that contains it.

TT43533