[OU-ADD] delivery_auto_refresh: Migration script to delete deactivated parameters
In v15 the module does not set the default parameters to any value as
was the case in v14. In v15 when a parameter is activated from settings
a record is created in the database and when it is deactivated the
record is deleted. For this reason and to avoid possible problems with
the configuration of the parameters, those that are deactivated are
deleted from the database.

TT43515